refactor(continuity): promote ContinuityState to astra-turn-types, validator hook, schema choke-point, redact, TDD coverage
- Move ContinuityState + SessionFacts to astra-turn-types (leaf crate, zero astra deps)
- astra-turn-core::continuity becomes a pure re-export (backward compat, zero call-site changes)
- HeavyCheckpoint.continuity_state: Option<serde_json::Value> with validate_with<F> hook
- step_restore::validate_restored_continuity_state wires hook to try_from_checkpoint_value
- ContinuityStateSchemaError.source is pub(crate), exposed via json_source() read-only
- ValidationError drops Serialize/Deserialize (no persistence contract)
- ensure_goal / ensure_tracked_goal redact secrets on intake
- No-tool-turn todo closes with synthetic evidence 'answered without tool invocation'
- todo id collision fixed: completed runtime-goal resets on new trackable goal
- execute_turn_strips_stale_manifest_and_injects_fresh_manifest_before_host_call test locks timing invariant
- sync_facts consistency covered by pure-state tests
- 21 new TDD tests across astra-turn-types, astra-pipeline, runtime
