+++ rust/crates/astra-cli/src/cli/chat_stream/sse_loop/agentic_sse_loop.rs
@@ -103,6 +103,7 @@ pub(crate) struct StreamResultBuild<'a> {
     pub(crate) step_recorder: &'a StepRecorder,
     pub(crate) turn_guard: &'a TurnGuard,
     pub(crate) last_heavy_checkpoint: Option<StepCheckpoint>,
+    pub(crate) runtime_continuity: astra_turn_types::continuity::ContinuityState,
     pub(crate) ttft_ms: Option<u64>,
     pub(crate) context_ms: Option<u64>,
     pub(crate) selector_strategy: Option<String>,
@@ -168,6 +169,7 @@ pub(crate) fn build_stream_result(ctx: StreamResultBuild<'_>) -> StreamResult {
         step_recorder,
         turn_guard,
         last_heavy_checkpoint,
+        runtime_continuity,
         ttft_ms,
         context_ms,
         selector_strategy,
@@ -234,6 +236,7 @@ pub(crate) fn build_stream_result(ctx: StreamResultBuild<'_>) -> StreamResult {
         step_recorder_summary: Some(step_recorder.summary()),
         tool_health_export: turn_guard.health.export_merged(tool_health_entries),
         last_heavy_checkpoint,
+        runtime_continuity,
         ttft_ms,
         context_ms,
         selector_strategy,
@@ -286,6 +289,7 @@ mod tests {
             step_recorder,
             turn_guard,
             last_heavy_checkpoint: None,
+            runtime_continuity: Default::default(),
             ttft_ms: Some(42),
             context_ms: Some(100),
             selector_strategy: Some("tfidf".into()),
