+++ rust/crates/astra-cli/src/cli/slash_session.rs
@@ -4522,6 +4522,7 @@ fn reset_state_for_session_restore(state: &mut ReplState) {
     state.last_delivery_report = None;
     state.redo_stack.clear();
     state.resume_guidance = None;
+    state.runtime_continuity = None;
     state.pending_goal_progress = None;
     state.pending_adaptive_state = None;
     state.pinned_skills.clear();
@@ -4635,23 +4636,43 @@ fn build_step_resume_guidance(
     })
 }
 
-fn build_continuity_resume_guidance(
-    continuity_state: Option<&serde_json::Value>,
-) -> Option<String> {
-    let value = continuity_state?;
-    let continuity = astra_turn_core::continuity::try_from_checkpoint_value(value)
-        .map_err(|e| {
-            tracing::warn!(
-                error = %e,
-                "dropped persisted continuity_state: schema mismatch"
-            );
-        })
-        .ok()?;
-    let manifest = astra_turn_core::continuity::AttentionManifest::from_state(&continuity, 4_000);
-    Some(format!(
+fn build_continuity_resume_guidance_from_state(
+    continuity: &astra_turn_types::continuity::ContinuityState,
+) -> String {
+    let manifest = astra_turn_types::continuity::AttentionManifest::from_state(&continuity, 4_000);
+    format!(
         "[Recovered runtime continuity]\nRestore this runtime-owned attention state before continuing. It is ground truth over narrative summaries.\n\n{}",
         manifest.as_str()
-    ))
+    )
+}
+
+fn parse_runtime_continuity_from_checkpoint(
+    continuity_state: Option<&serde_json::Value>,
+    source: &'static str,
+) -> Option<astra_turn_types::continuity::ContinuityState> {
+    let Some(value) = continuity_state else {
+        return None;
+    };
+    astra_turn_types::continuity::try_from_checkpoint_value(value)
+        .map_err(|error| {
+            tracing::warn!(
+                error = %error,
+                source,
+                "dropping invalid continuity_state"
+            );
+        })
+        .ok()
+}
+
+fn restore_runtime_continuity_guidance(
+    state: &mut ReplState,
+    continuity_state: Option<&serde_json::Value>,
+    source: &'static str,
+) -> Option<String> {
+    let continuity = parse_runtime_continuity_from_checkpoint(continuity_state, source)?;
+    let guidance = build_continuity_resume_guidance_from_state(&continuity);
+    state.runtime_continuity = Some(continuity);
+    Some(guidance)
 }
 
 fn history_pairs_from_messages(messages: &[serde_json::Value]) -> Vec<(String, String)> {
@@ -4773,9 +4794,31 @@ async fn apply_restored_session(
 
     apply_restored_workspace_state(state, &restored.session_id);
 
-    if let Ok(Some(step_restored)) =
-        astra_runtime::pipeline::step_restore::restore_session(&restored.session_id)
-    {
+    let step_restored =
+        match astra_runtime::pipeline::step_restore::restore_session_with_continuity_validator(
+            &restored.session_id,
+            |value| {
+                astra_turn_types::continuity::try_from_checkpoint_value(value)
+                    .map(|_| ())
+                    .map_err(|e| e.to_string())
+            },
+        ) {
+            Ok(restored) => restored,
+            Err(astra_runtime::pipeline::step_restore::RestoreError::IoError(error)) => {
+                return Err(format!(
+                    "Failed to read local step checkpoint for {}: {}",
+                    restored.session_id, error
+                ));
+            }
+            Err(error) => {
+                tracing::warn!(
+                    error = %error,
+                    "skipping invalid local step checkpoint during resume"
+                );
+                None
+            }
+        };
+    if let Some(step_restored) = step_restored {
         let summary = astra_runtime::pipeline::step_restore::restore_summary(&step_restored);
         for tool in &step_restored.blocked_tools {
             if !state.tool_health_entries.iter().any(|e| e.name == *tool) {
@@ -4793,7 +4836,11 @@ async fn apply_restored_session(
         );
         let step_guidance = combine_resume_guidance(
             step_guidance,
-            build_continuity_resume_guidance(step_restored.continuity_state.as_ref()),
+            restore_runtime_continuity_guidance(
+                state,
+                step_restored.continuity_state.as_ref(),
+                "local step checkpoint",
+            ),
         );
         state.resume_guidance = combine_resume_guidance(
             step_guidance,
@@ -4812,7 +4859,11 @@ async fn apply_restored_session(
                 heavy.interruption.as_ref(),
                 heavy.compaction_state.as_ref(),
             ),
-            build_continuity_resume_guidance(heavy.continuity_state.as_ref()),
+            restore_runtime_continuity_guidance(
+                state,
+                heavy.continuity_state.as_ref(),
+                "local heavy checkpoint",
+            ),
         );
         state.resume_guidance = combine_resume_guidance(
             step_guidance,
@@ -4831,7 +4882,11 @@ async fn apply_restored_session(
                 restored.interruption.as_ref(),
                 restored.compaction_state.as_ref(),
             ),
-            build_continuity_resume_guidance(restored.continuity_state.as_ref()),
+            restore_runtime_continuity_guidance(
+                state,
+                restored.continuity_state.as_ref(),
+                "restored cloud checkpoint",
+            ),
         );
         state.resume_guidance = combine_resume_guidance(
             step_guidance,
@@ -4850,7 +4905,7 @@ async fn apply_restored_session(
                 match astra_services::session_restore::parse_cloud_heavy_checkpoint_state(
                     &state_json,
                 ) {
-                    Some(heavy) => {
+                    Ok(Some(heavy)) => {
                         apply_heavy_state_fallback(
                             state,
                             &heavy.blocked_tools,
@@ -4863,7 +4918,11 @@ async fn apply_restored_session(
                                 heavy.interruption.as_ref(),
                                 heavy.compaction_state.as_ref(),
                             ),
-                            build_continuity_resume_guidance(heavy.continuity_state.as_ref()),
+                            restore_runtime_continuity_guidance(
+                                state,
+                                heavy.continuity_state.as_ref(),
+                                "pulled cloud checkpoint",
+                            ),
                         );
                         state.resume_guidance = combine_resume_guidance(
                             step_guidance,
@@ -4871,10 +4930,16 @@ async fn apply_restored_session(
                         );
                         eprintln!("  {} Restored step checkpoint from cloud", "☁".cyan());
                     }
-                    None => {
+                    Ok(None) => {
                         eprintln!(
                             "  {} Cloud checkpoint corrupted, skipping",
                             theme::icon_warn()
+                        );
+                    }
+                    Err(e) => {
+                        tracing::warn!(
+                            error = %e,
+                            "skipping cloud step checkpoint with invalid continuity_state"
                         );
                     }
                 }
@@ -5510,6 +5575,33 @@ mod resume_tests {
         std::fs::create_dir_all(&sessions).unwrap();
         let guard = JournalDirGuard::new(&sessions);
         (tmp, guard)
+    }
+
+    #[test]
+    fn restore_runtime_continuity_guidance_sets_repl_state() {
+        let mut continuity = astra_turn_types::continuity::ContinuityState::default();
+        continuity.ensure_tracked_goal("continue the checkpoint restore implementation");
+        let value = serde_json::to_value(&continuity).unwrap();
+        let mut state = ReplState::default();
+
+        let guidance =
+            restore_runtime_continuity_guidance(&mut state, Some(&value), "test checkpoint")
+                .expect("guidance");
+
+        assert!(guidance.contains("[Recovered runtime continuity]"));
+        assert_eq!(state.runtime_continuity, Some(continuity));
+    }
+
+    #[test]
+    fn restore_runtime_continuity_guidance_soft_drops_invalid_state() {
+        let mut state = ReplState::default();
+        let invalid = serde_json::json!({});
+
+        let guidance =
+            restore_runtime_continuity_guidance(&mut state, Some(&invalid), "test checkpoint");
+
+        assert!(guidance.is_none());
+        assert!(state.runtime_continuity.is_none());
     }
 
     fn write_local_resumable_session(session_id: &str, turn_count: u32) {
