+++ rust/crates/astra-cli/src/cli/slash_state.rs
@@ -368,6 +368,7 @@ pub(super) async fn handle_state_command(
                     git_worktree_journal: None,
                     session_state_journal: None,
                     task_manager: None,
+                    runtime_continuity: None,
                     turn_index: 0,
                     evolution_service: state.evolution_service.clone(),
                 }) => r,
@@ -475,6 +476,7 @@ pub(super) async fn handle_state_command(
                             git_worktree_journal: None,
                             session_state_journal: None,
                             task_manager: None,
+                            runtime_continuity: None,
                             turn_index: 0,
                             evolution_service: state.evolution_service.clone(),
                         })
@@ -563,6 +565,7 @@ pub(super) async fn handle_state_command(
                                     git_worktree_journal: None,
                                     session_state_journal: None,
                                     task_manager: None,
+                                    runtime_continuity: None,
                                     turn_index: 0,
                                     evolution_service: state.evolution_service.clone(),
                                 })
