+++ rust/crates/astra-cli/src/cli/slash_task.rs
@@ -313,6 +313,7 @@ pub(super) async fn handle_task_command(
                     git_worktree_journal: None,
                     session_state_journal: None,
                     task_manager: None,
+                    runtime_continuity: None,
                     turn_index: 0,
                     evolution_service: bg_evolution_service.clone(),
                 })
