+++ rust/crates/astra-cli/src/main.rs
@@ -1497,6 +1497,7 @@ mod tests {
             git_worktree_journal: None,
             session_state_journal: None,
             task_manager: None,
+            runtime_continuity: None,
             turn_index: 0,
             evolution_service: None,
         })
@@ -1569,6 +1570,7 @@ mod tests {
             git_worktree_journal: None,
             session_state_journal: None,
             task_manager: None,
+            runtime_continuity: None,
             turn_index: 0,
             evolution_service: None,
         })
@@ -1657,6 +1659,7 @@ mod tests {
             git_worktree_journal: None,
             session_state_journal: None,
             task_manager: None,
+            runtime_continuity: None,
             turn_index: 0,
             evolution_service: None,
         })
