+++ rust/crates/astra-turn-core/src/cloud_session_facts.rs
@@ -1,322 +1,87 @@
-//! L1a: System-tracked session facts (ground truth, zero LLM).
+//! Journal ingestion adapter for shared session facts.
 //!
-//! Updated every turn from journal events and checkpoint state.
-//! See `docs/design/session-memory-protocol.md` Section 4.1.
+//! The pure `SessionFacts` data model lives in `astra-turn-types`; this module
+//! keeps service-layer journal record ingestion in `astra-turn-core` so the type
+//! crate remains independent of restore/journal services.
+
+pub use astra_turn_types::session_facts::*;
 
 use astra_services::session_journal::{JournalEvent, ToolCallRecord};
-use serde::{Deserialize, Serialize};
-use std::fmt::Write;
-
-// ── Types ────────────────────────────────────────────────────────────────────
-
-/// Ground truth session state. Never hallucinated.
-#[derive(Debug, Clone, Default, Serialize, Deserialize, PartialEq, Eq)]
-pub struct SessionFacts {
-    /// Files touched this session, most recent last. Capped at 20.
-    pub active_files: Vec<FileEntry>,
-    /// Last N tool calls with outcomes. Capped at 10.
-    pub recent_tool_calls: Vec<ToolFact>,
-    /// Plan progress (from checkpoint, not journal).
-    pub plan_state: Option<PlanFact>,
-    /// Blocked/unhealthy tools (from checkpoint).
-    pub blocked_tools: Vec<String>,
-    /// Error accumulator.
-    pub error_state: ErrorFact,
-    /// Current turn number.
-    pub turn: u32,
-    /// Cumulative prompt tokens.
-    pub estimated_tokens: u64,
-}
-
-#[derive(Debug, Clone, Serialize, Deserialize, PartialEq, Eq)]
-pub struct FileEntry {
-    pub path: String,
-    /// "read", "write", or "create"; "write" covers all non-create mutations, including deletes.
-    pub last_action: String,
-    pub turn: u32,
-}
-
-#[derive(Debug, Clone, Serialize, Deserialize, PartialEq, Eq)]
-pub struct ToolFact {
-    pub name: String,
-    pub ok: bool,
-    pub turn: u32,
-}
-
-#[derive(Debug, Clone, Default, Serialize, Deserialize, PartialEq, Eq)]
-pub struct PlanFact {
-    pub goal: String,
-    pub completed: u32,
-    pub total: u32,
-    pub current_subtask: Option<String>,
-}
-
-#[derive(Debug, Clone, Default, Serialize, Deserialize, PartialEq, Eq)]
-pub struct ErrorFact {
-    pub total_errors: u32,
-    pub last_error: Option<String>,
-    pub last_error_turn: Option<u32>,
-}
 
 const MAX_ACTIVE_FILES: usize = 20;
 const MAX_RECENT_TOOLS: usize = 10;
 
-// ── Update ───────────────────────────────────────────────────────────────────
+/// Incremental update from a single turn's journal event.
+pub fn update_from_journal_event(facts: &mut SessionFacts, event: &JournalEvent) {
+    if let Some(t) = event.turn {
+        facts.turn = t;
+    }
+    if let Some(tokens) = event.tokens_in {
+        facts.estimated_tokens += tokens;
+    }
 
-impl SessionFacts {
-    /// Incremental update from a single turn's journal event.
-    pub fn update_from_journal_event(&mut self, event: &JournalEvent) {
-        if let Some(t) = event.turn {
-            self.turn = t;
-        }
-        if let Some(tokens) = event.tokens_in {
-            self.estimated_tokens += tokens;
-        }
-
-        // Extract file paths and tool outcomes from tool_calls
-        if let Some(tool_calls) = &event.tool_calls {
-            for tc in tool_calls {
-                if tc.is_synthetic_placeholder() {
-                    continue;
-                }
-                // File tracking
-                if let Some(path) = extract_file_path(tc) {
-                    let action = action_for_tool(&tc.name);
-                    self.upsert_file(path, action, self.turn);
-                }
-                // Tool outcome tracking
-                self.recent_tool_calls.push(ToolFact {
-                    name: tc.name.clone(),
-                    ok: tc.ok,
-                    turn: self.turn,
-                });
-                if self.recent_tool_calls.len() > MAX_RECENT_TOOLS {
-                    self.recent_tool_calls.remove(0);
-                }
+    if let Some(tool_calls) = &event.tool_calls {
+        for tc in tool_calls {
+            if tc.is_synthetic_placeholder() {
+                continue;
             }
-        }
-
-        // Error tracking
-        if let Some(err) = &event.error {
-            self.error_state.total_errors += 1;
-            self.error_state.last_error = Some(truncate(err, 200));
-            self.error_state.last_error_turn = Some(self.turn);
-        }
-    }
-
-    /// Set blocked tools from checkpoint state.
-    pub fn set_blocked_tools(&mut self, blocked: Vec<String>) {
-        self.blocked_tools = blocked;
-    }
-
-    /// Set plan state from checkpoint's `executing_plan_json`.
-    pub fn set_plan_state(&mut self, plan: Option<PlanFact>) {
-        self.plan_state = plan;
-    }
-
-    fn upsert_file(&mut self, path: String, action: String, turn: u32) {
-        // Update existing entry or append
-        if let Some(entry) = self.active_files.iter_mut().find(|f| f.path == path) {
-            entry.last_action = action;
-            entry.turn = turn;
-        } else {
-            self.active_files.push(FileEntry {
-                path,
-                last_action: action,
-                turn,
+            if let Some(path) = extract_file_path(tc) {
+                upsert_file(facts, path, action_for_tool(&tc.name), facts.turn);
+            }
+            facts.recent_tool_calls.push(ToolFact {
+                name: tc.name.clone(),
+                ok: tc.ok,
+                turn: facts.turn,
             });
-            if self.active_files.len() > MAX_ACTIVE_FILES {
-                self.active_files.remove(0);
+            if facts.recent_tool_calls.len() > MAX_RECENT_TOOLS {
+                facts.recent_tool_calls.remove(0);
             }
         }
+    }
+
+    if let Some(err) = &event.error {
+        facts.error_state.total_errors += 1;
+        facts.error_state.last_error = Some(truncate(err, 200));
+        facts.error_state.last_error_turn = Some(facts.turn);
     }
 }
 
-// ── Injection ────────────────────────────────────────────────────────────────
-
-impl SessionFacts {
-    /// Deterministic working-set injection for cross-turn continuity.
-    ///
-    /// Field order is stable by design so prefix-cache providers can reuse the
-    /// surrounding prompt while still preserving the facts the model needs to
-    /// stay oriented after compaction.
-    pub fn to_working_set_injection(&self, current_goal: &str) -> String {
-        let mut out = String::from("[working-set:v1]\n");
-        let goal = if let Some(plan) = &self.plan_state {
-            plan.goal.trim()
-        } else {
-            current_goal.trim()
-        };
-        writeln!(out, "goal: {}", truncate_or_none(goal, 240)).ok();
-
-        let pending = self
-            .plan_state
-            .as_ref()
-            .and_then(|plan| plan.current_subtask.as_deref())
-            .map(str::trim)
-            .filter(|value| !value.is_empty())
-            .map(|value| truncate(value, 200))
-            .unwrap_or_else(|| "none".to_string());
-        writeln!(out, "pending_work: {pending}").ok();
-
-        out.push_str("active_files:\n");
-        if self.active_files.is_empty() {
-            out.push_str("- none\n");
-        } else {
-            let mut files: Vec<&FileEntry> = self.active_files.iter().collect();
-            files.sort_by(|a, b| a.path.cmp(&b.path));
-            for file in files.into_iter().take(12) {
-                writeln!(
-                    out,
-                    "- {} [{} t{}]",
-                    truncate(&file.path, 160),
-                    file.last_action,
-                    file.turn
-                )
-                .ok();
-            }
+fn upsert_file(facts: &mut SessionFacts, path: String, action: String, turn: u32) {
+    if let Some(entry) = facts.active_files.iter_mut().find(|f| f.path == path) {
+        entry.last_action = action;
+        entry.turn = turn;
+    } else {
+        facts.active_files.push(FileEntry {
+            path,
+            last_action: action,
+            turn,
+        });
+        if facts.active_files.len() > MAX_ACTIVE_FILES {
+            facts.active_files.remove(0);
         }
-
-        out.push_str("recent_tools:\n");
-        if self.recent_tool_calls.is_empty() {
-            out.push_str("- none\n");
-        } else {
-            for tool in self.recent_tool_calls.iter().rev().take(6).rev() {
-                writeln!(
-                    out,
-                    "- {} [{} t{}]",
-                    tool.name,
-                    if tool.ok { "ok" } else { "error" },
-                    tool.turn
-                )
-                .ok();
-            }
-        }
-
-        out.push_str("tool_risks:\n");
-        if self.blocked_tools.is_empty() && self.error_state.total_errors == 0 {
-            out.push_str("- none\n");
-        } else {
-            if !self.blocked_tools.is_empty() {
-                let mut blocked = self.blocked_tools.clone();
-                blocked.sort();
-                writeln!(out, "- blocked: {}", blocked.join(", ")).ok();
-            }
-            if self.error_state.total_errors > 0 {
-                let last = self
-                    .error_state
-                    .last_error
-                    .as_deref()
-                    .map(|err| truncate(err, 180))
-                    .unwrap_or_else(|| "unknown".to_string());
-                writeln!(
-                    out,
-                    "- errors: {} total, last: {}",
-                    self.error_state.total_errors, last
-                )
-                .ok();
-            }
-        }
-
-        out
-    }
-
-    /// Serialize to injection format (~150 tokens).
-    pub fn to_injection(&self) -> String {
-        let mut out = String::from("# System State\n");
-        writeln!(
-            out,
-            "Turn {}, ~{}K tokens",
-            self.turn,
-            self.estimated_tokens / 1000
-        )
-        .ok();
-
-        if let Some(plan) = &self.plan_state {
-            write!(
-                out,
-                "Plan: {} ({}/{})",
-                plan.goal, plan.completed, plan.total
-            )
-            .ok();
-            if let Some(sub) = &plan.current_subtask {
-                write!(out, ", current: {sub}").ok();
-            }
-            out.push('\n');
-        }
-
-        if !self.active_files.is_empty() {
-            out.push_str("Active files:\n");
-            for f in self.active_files.iter().rev().take(10) {
-                writeln!(out, "  {} {} (t{})", f.last_action, f.path, f.turn).ok();
-            }
-        }
-
-        if self.error_state.total_errors > 0 {
-            write!(out, "Errors: {} total", self.error_state.total_errors).ok();
-            if let Some(err) = &self.error_state.last_error {
-                write!(out, ", last: {err}").ok();
-            }
-            out.push('\n');
-        }
-
-        if !self.blocked_tools.is_empty() {
-            writeln!(out, "Blocked tools: {}", self.blocked_tools.join(", ")).ok();
-        }
-
-        out
-    }
-
-    /// Check if a file path is in the active set (for compaction pin list).
-    pub fn is_active_file(&self, path: &str, recent_turns: u32) -> bool {
-        let cutoff = self.turn.saturating_sub(recent_turns);
-        self.active_files
-            .iter()
-            .any(|f| f.path == path && f.turn >= cutoff)
-    }
-
-    /// Check whether a file path is explicitly referenced by pending plan work.
-    pub fn is_pending_relevant_file(&self, path: &str) -> bool {
-        let Some(plan) = &self.plan_state else {
-            return false;
-        };
-        let Some(subtask) = plan.current_subtask.as_deref() else {
-            return false;
-        };
-        if path.is_empty() {
-            return false;
-        }
-        subtask.contains(path)
     }
 }
-
-// ── Helpers ──────────────────────────────────────────────────────────────────
 
 /// Extract file path from a ToolCallRecord.
 /// Uses `file_path` field if available, falls back to parsing `args_full` (untruncated)
 /// and finally `args_preview` (which may be truncated mid-path).
 fn extract_file_path(tc: &ToolCallRecord) -> Option<String> {
-    // Prefer the dedicated field (populated at tool execution time)
-    if let Some(fp) = &tc.file_path {
-        if !fp.is_empty() {
-            return Some(fp.clone());
-        }
+    if let Some(fp) = &tc.file_path
+        && !fp.is_empty()
+    {
+        return Some(fp.clone());
     }
-    // Next-best: parse the untruncated args_full. Reliable for str_replace and
-    // any record where `args_preview` would have been cut off mid-path.
-    if let Some(full) = tc.args_full.as_deref() {
-        if let Some(path) = parse_path_from_json_preview(full) {
-            return Some(path);
-        }
+    if let Some(full) = tc.args_full.as_deref()
+        && let Some(path) = parse_path_from_json_preview(full)
+    {
+        return Some(path);
     }
-    // Last-resort: parse args_preview. Best-effort for legacy records.
     let preview = tc.args_preview.as_deref()?;
     parse_path_from_json_preview(preview)
 }
 
 /// Best-effort extraction of "path" field from a truncated JSON preview.
 fn parse_path_from_json_preview(preview: &str) -> Option<String> {
-    // Look for "path":"..." or "path": "..."
     let idx = preview.find("\"path\"")?;
     let rest = &preview[idx + 6..];
     let colon = rest.find(':')?;
@@ -351,14 +116,6 @@ fn truncate(s: &str, max: usize) -> String {
     } else {
         let boundary = s.floor_char_boundary(max);
         format!("{}…", &s[..boundary])
-    }
-}
-
-fn truncate_or_none(s: &str, max: usize) -> String {
-    if s.is_empty() {
-        "none".to_string()
-    } else {
-        truncate(s, max)
     }
 }
 
@@ -409,7 +166,7 @@ mod tests {
                 make_tc("str_replace", true, Some("src/lib.rs"), None),
             ],
         );
-        facts.update_from_journal_event(&event);
+        update_from_journal_event(&mut facts, &event);
 
         assert_eq!(facts.active_files.len(), 2);
         assert_eq!(facts.active_files[0].path, "src/main.rs");
@@ -430,7 +187,7 @@ mod tests {
                 make_tc("delete_file", true, Some("gone.rs"), None),
             ],
         );
-        facts.update_from_journal_event(&event);
+        update_from_journal_event(&mut facts, &event);
 
         assert_eq!(facts.active_files.len(), 4);
         assert_eq!(facts.active_files[0].last_action, "create");
@@ -451,7 +208,7 @@ mod tests {
                 Some(r#"{"path":"src/foo.rs"}"#),
             )],
         );
-        facts.update_from_journal_event(&event);
+        update_from_journal_event(&mut facts, &event);
 
         assert_eq!(facts.active_files.len(), 1);
         assert_eq!(facts.active_files[0].path, "src/foo.rs");
@@ -469,7 +226,7 @@ mod tests {
         let mut tc = make_tc("str_replace", true, None, Some(truncated_preview));
         tc.args_full = Some(full.to_string());
         let event = make_event(1, vec![tc]);
-        facts.update_from_journal_event(&event);
+        update_from_journal_event(&mut facts, &event);
 
         assert_eq!(facts.active_files.len(), 1);
         assert_eq!(
@@ -484,8 +241,8 @@ mod tests {
         let mut facts = SessionFacts::default();
         let e1 = make_event(1, vec![make_tc("read_file", true, Some("a.rs"), None)]);
         let e2 = make_event(2, vec![make_tc("str_replace", true, Some("a.rs"), None)]);
-        facts.update_from_journal_event(&e1);
-        facts.update_from_journal_event(&e2);
+        update_from_journal_event(&mut facts, &e1);
+        update_from_journal_event(&mut facts, &e2);
 
         assert_eq!(facts.active_files.len(), 1);
         assert_eq!(facts.active_files[0].last_action, "write");
@@ -505,7 +262,7 @@ mod tests {
                     None,
                 )],
             );
-            facts.update_from_journal_event(&event);
+            update_from_journal_event(&mut facts, &event);
         }
         assert_eq!(facts.active_files.len(), MAX_ACTIVE_FILES);
         // Oldest should be dropped
@@ -517,7 +274,7 @@ mod tests {
         let mut facts = SessionFacts::default();
         let mut event = make_event(3, vec![]);
         event.error = Some("sqlx migration failed".to_string());
-        facts.update_from_journal_event(&event);
+        update_from_journal_event(&mut facts, &event);
 
         assert_eq!(facts.error_state.total_errors, 1);
         assert_eq!(
@@ -537,7 +294,7 @@ mod tests {
                 make_tc("bash", false, None, None),
             ],
         );
-        facts.update_from_journal_event(&event);
+        update_from_journal_event(&mut facts, &event);
 
         assert_eq!(facts.recent_tool_calls.len(), 2);
         assert!(facts.recent_tool_calls[0].ok);
@@ -549,7 +306,7 @@ mod tests {
         let mut facts = SessionFacts::default();
         for i in 0..15 {
             let event = make_event(i, vec![make_tc("bash", true, None, None)]);
-            facts.update_from_journal_event(&event);
+            update_from_journal_event(&mut facts, &event);
         }
         assert_eq!(facts.recent_tool_calls.len(), MAX_RECENT_TOOLS);
     }
@@ -559,8 +316,8 @@ mod tests {
         let mut facts = SessionFacts::default();
         let e1 = make_event(1, vec![make_tc("read_file", true, Some("old.rs"), None)]);
         let e2 = make_event(10, vec![make_tc("read_file", true, Some("new.rs"), None)]);
-        facts.update_from_journal_event(&e1);
-        facts.update_from_journal_event(&e2);
+        update_from_journal_event(&mut facts, &e1);
+        update_from_journal_event(&mut facts, &e2);
 
         assert!(facts.is_active_file("new.rs", 5));
         assert!(!facts.is_active_file("old.rs", 5));
@@ -704,8 +461,8 @@ mod tests {
         let mut facts = SessionFacts::default();
         let e1 = make_event(1, vec![]);
         let e2 = make_event(2, vec![]);
-        facts.update_from_journal_event(&e1);
-        facts.update_from_journal_event(&e2);
+        update_from_journal_event(&mut facts, &e1);
+        update_from_journal_event(&mut facts, &e2);
         assert_eq!(facts.estimated_tokens, 2000);
     }
 }
