+++ rust/crates/astra-turn-types/Cargo.toml
@@ -8,3 +8,4 @@ description = "Core turn types for astra runtime"
 regex = "1"
 serde = { version = "1", features = ["derive"] }
 serde_json = "1"
+thiserror = { workspace = true }
