+++ rust/crates/runtime/src/turn/agentic_loop_host.rs
@@ -757,7 +757,7 @@ pub struct AgenticLoopState {
     // ── Runtime-owned continuity (goal/todo/attention) ──
     /// Deterministic continuity state used to reconstruct attention every LLM round
     /// without relying on model-invoked task tools.
-    pub continuity: astra_turn_core::continuity::ContinuityState,
+    pub continuity: astra_turn_types::continuity::ContinuityState,
 
     /// Provider-aware compaction strategy for microcompact placeholders.
     pub compact_strategy: astra_turn_core::microcompact::CompactStrategy,
@@ -1167,6 +1167,7 @@ pub(crate) mod tests {
         reflection_error: Option<String>,
         pub(crate) last_reflection_prompt: Option<String>,
         pub(crate) rendered_final_text: Vec<String>,
+        pub(crate) executed_messages: Vec<Vec<Value>>,
     }
 
     impl MockHost {
@@ -1182,6 +1183,7 @@ pub(crate) mod tests {
                 reflection_error: None,
                 last_reflection_prompt: None,
                 rendered_final_text: Vec::new(),
+                executed_messages: Vec::new(),
             }
         }
 
@@ -1209,7 +1211,7 @@ pub(crate) mod tests {
     impl AgenticLoopHost for MockHost {
         async fn execute_turn(
             &mut self,
-            _state: &mut AgenticLoopState,
+            state: &mut AgenticLoopState,
         ) -> Result<HostTurnResult, astra_core::ClassifiedError> {
             if self.turn_results.is_empty() {
                 return Err(astra_core::ClassifiedError::new(
@@ -1217,6 +1219,7 @@ pub(crate) mod tests {
                     "no more turns",
                 ));
             }
+            self.executed_messages.push(state.messages.clone());
             let result = self.turn_results.remove(0);
             self.current_turn += 1;
             Ok(result)
