+++ rust/crates/runtime/src/turn/cloud/session_facts.rs
@@ -1,1 +1,2 @@
-pub use astra_turn_core::cloud_session_facts::*;
+pub use astra_turn_core::cloud_session_facts::update_from_journal_event;
+pub use astra_turn_types::session_facts::*;
