+++ rust/crates/services/Cargo.toml
@@ -6,6 +6,7 @@ edition.workspace = true
 [dependencies]
 astra-core.workspace = true
 astra-logging.workspace = true
+astra-turn-types.workspace = true
 async-trait.workspace = true
 axum.workspace = true
 base64.workspace = true
